+++ src/UriTemplates/Properties/AssemblyInfo.cs
@@ -19,8 +19,6 @@ using System.Runtime.InteropServices;
 // COM, set the ComVisible attribute to true on that type.
 [assembly: ComVisible(false)]
 
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("8cdd9647-dfb9-4560-8428-598dc8ea564f")]
 
 // Version information for an assembly consists of the following four values:
 //